python-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
numpy==1.26.4
opencv-python-headless==4.9.0.80
insightface==0.7.3
onnxruntime==1.18.0
Pillow==10.3.0
scikit-learn==1.4.2
httpx==0.27.0
